+++ src/app/immer-app/components/todo-description.component.spec.ts
@@ -8,30 +8,34 @@ import { ImmerDataStrategy } from '@ng-state/immer-data-strategy';
 describe('TodoDescription', () => {
 
     let component: TodoDescription;
-    let copyIntitialState: typeof initialState;
+    let copyInitialState: typeof initialState;
     const cd = { markForCheck: () => { } };
 
     beforeEach(() => {
         NgStateTestBed.setTestEnvironment(new ImmerDataStrategy());
         NgStateTestBed.strictActionsCheck = false;
 
-        copyIntitialState = JSON.parse(JSON.stringify(initialState));
-        copyIntitialState.todos.push(<TodoModel>{ description: 'test description' });
-        NgStateTestBed.createActions(TodoStateActions, copyIntitialState, ['todos', 1]) as TodoStateActions;
+        copyInitialState = JSON.parse(JSON.stringify(initialState));
+        copyInitialState.todos.push(<TodoModel>{ description: 'test description' });
+        NgStateTestBed.createStore(copyInitialState);
         component = new TodoDescription(cd as any);
-        component.ngOnInit();
     });
 
     it('should get description - immer', () => {
+        component.stateIndex = 1;
+        component.statePath = ['todos'];
+        component.ngOnInit();
         expect(component.actions.todoDescription).toEqual('test description');
     });
 
     it('should set actions to component - immer', () => {
-        copyIntitialState.todos.push(<TodoModel>{ description: 'test description 2' });
+        copyInitialState.todos.push(<TodoModel>{ description: 'test description 2' });
 
-        const actions = NgStateTestBed.createActions(TodoStateActions, copyIntitialState, ['todos', 1]) as TodoStateActions;
-        NgStateTestBed.setActionsToComponent(actions, component);
+        component.stateIndex = 2;
+        component.statePath = ['todos'];
+        NgStateTestBed.createStore(copyInitialState);
+        component.ngOnInit();
 
-        expect(component.actions.todoDescription).toEqual('test description');
+        expect(component.actions.todoDescription).toEqual('test description 2');
     });
 });